+++ gs/internal/gs_core/bean.go
@@ -23,9 +23,9 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/go-spring/spring-core/conf"
 	"github.com/go-spring/spring-core/gs/internal/gs"
 	"github.com/go-spring/spring-core/gs/internal/gs_arg"
+	"github.com/go-spring/spring-core/util"
 )
 
 // NewBean 普通函数注册时需要使用 reflect.ValueOf(fn) 形式以避免和构造函数发生冲突。
@@ -95,7 +95,7 @@ func NewBean(objOrCtor interface{}, ctorArgs ...gs.Arg) *gs.BeanDefinition {
 		method = strings.LastIndexByte(fnInfo.Name(), ')') > 0
 	}
 
-	if t.Kind() == reflect.Ptr && !conf.IsValueType(t.Elem()) {
+	if t.Kind() == reflect.Ptr && !util.IsValueType(t.Elem()) {
 		panic(errors.New("bean should be *val but not *ref"))
 	}
